squash and rebase andronov-alexey:treat_warnings_as_errors

+++ src/lib/component/controller/ErrorController.cpp
@@ -73,7 +73,7 @@ void ErrorController::handleMessage(MessageErrorCountUpdate* message)
 
 	ErrorFilter filter = getView()->getErrorFilter();
 
-	int room = filter.limit - m_errorCount;
+	int room = static_cast<int>(filter.limit) - static_cast<int>(m_errorCount);
 	if (room > 0)
 	{
 		filter.limit = 0;
